+++ src/plugins/intel_cpu/src/nodes/kernels/aarch64/jit_uni_eltwise_generic.hpp
@@ -28,7 +28,7 @@
 
 #include "emitters/plugin/aarch64/jit_eltwise_emitters.hpp"
 #include "emitters/plugin/aarch64/jit_emitter.hpp"
-#include "nodes/kernels/jit_eltwise_call_args_ptrs.hpp"
+#include "nodes/kernels/jit_eltwise_common.hpp"
 #include "utils/cpu_utils.hpp"
 #include "utils/general_utils.h"
 
@@ -40,45 +40,6 @@ using namespace Xbyak_aarch64;
 using namespace dnnl::impl::cpu;
 using namespace dnnl::impl::cpu::aarch64;
 
-struct jit_eltwise_params {
-    size_t inputs_number;
-    size_t input_size;
-
-    ov::element::Type src_prc[MAX_ELTWISE_INPUTS];
-    ov::element::Type dst_prc;
-
-    VectorDims dims;
-    VectorDims src_offsets[MAX_ELTWISE_INPUTS];
-    VectorDims dst_offsets;
-    VectorDims oc_offsets;
-
-    size_t src_size[MAX_ELTWISE_INPUTS];
-    size_t dst_size;
-    size_t oc_size;
-
-    size_t work_amount;
-    bool use_runtime_ptrs;
-    bool do_output_saturation;
-};
-
-struct jit_eltwise_call_args_indexes {
-    size_t indexes[MAX_ELTWISE_DIM_RANK];
-};
-
-struct jit_uni_eltwise_kernel {
-    void (*ker_)(const node::jit_eltwise_call_args_ptrs*, const jit_eltwise_call_args_indexes*);
-
-    void operator()(const node::jit_eltwise_call_args_ptrs* const_args, const jit_eltwise_call_args_indexes* indexes);
-
-    jit_uni_eltwise_kernel() {}
-    jit_uni_eltwise_kernel(jit_eltwise_params jep) : ker_(nullptr), jep_(std::move(jep)) {}
-    virtual ~jit_uni_eltwise_kernel() {}
-
-    virtual void create_ker() = 0;
-
-    jit_eltwise_params jep_;
-};
-
 template <dnnl::impl::cpu::aarch64::cpu_isa_t isa>
 struct jit_uni_eltwise_generic : public jit_uni_eltwise_kernel, jit_generator {
 public:
@@ -88,8 +49,6 @@ public:
                             std::vector<EltwiseData> eltwise_data,
                             std::vector<ov::intel_cpu::Type> ops_list,
                             dnnl::post_ops post_ops);
-
-    jit_uni_eltwise_generic() {}
 
     void create_ker() override {
         jit_generator::create_kernel();
@@ -253,16 +212,6 @@ private:
 
     std::shared_ptr<jit_emitter> eltwise_emitter = nullptr;
     std::vector<std::shared_ptr<jit_emitter>> post_op_emitters;
-};
-
-class eltwise_precision_helper {
-public:
-    static ov::element::Type get_precision(const size_t inputs_number,
-                                           const ov::element::Type (&src_prc)[MAX_ELTWISE_INPUTS],
-                                           const std::vector<EltwiseData>& eltwise_data);
-
-private:
-    static std::set<std::vector<element::Type>> get_supported_precisions(const Algorithm& algo);
 };
 
 }  // namespace aarch64
