+++ src/plugins/intel_cpu/tests/unit/nodes/eltwise_node_test.cpp
@@ -10,6 +10,11 @@ using namespace ov::intel_cpu;
 
 class EltwisePrecisionHelperTest : public testing::Test {};
 
+namespace EltwisePrecisionHelper {
+static const std::vector<ov::element::Type> exec_precisions_priority =
+        {ov::element::u8, ov::element::bf16, ov::element::i32, ov::element::f32};
+}  // namespace EltwisePrecisionHelper
+
 TEST(EltwisePrecisionHelperTest, get_precision_mixed) {
     ov::element::Type src_prc[MAX_ELTWISE_INPUTS];
     const size_t inputs_size = 4ull;
@@ -22,7 +27,8 @@ TEST(EltwisePrecisionHelperTest, get_precision_mixed) {
         {Algorithm::EltwiseMulAdd}
     };
 
-    const auto precision = ov::intel_cpu::node::eltwise_precision_helper::get_precision(inputs_size, src_prc, eltwise_data);
+    const auto precision =
+        ov::intel_cpu::eltwise_precision_helper::get_precision(inputs_size, src_prc, eltwise_data, EltwisePrecisionHelper::exec_precisions_priority);
     ASSERT_EQ(ov::element::i32, precision);
 }
 
@@ -38,6 +44,7 @@ TEST(EltwisePrecisionHelperTest, get_precision_single) {
         {Algorithm::EltwiseMod}
     };
 
-    const auto precision = ov::intel_cpu::node::eltwise_precision_helper::get_precision(inputs_size, src_prc, eltwise_data);
+    const auto precision =
+        ov::intel_cpu::eltwise_precision_helper::get_precision(inputs_size, src_prc, eltwise_data, EltwisePrecisionHelper::exec_precisions_priority);
     ASSERT_EQ(ov::element::f32, precision);
 }
